drivers: nfc: Rework Drivers over NXP Release

+++ drivers/nfc/nxp/ese/pn8xt.h
@@ -20,15 +20,15 @@
  * level 1 = Enable power
  * level 0 = Disable power
  */
-#define PN8XT_ESE_SET_SPI_PWR           _IOW(NXP_ESE_MAGIC, 0x04, unsigned int)
+#define PN8XT_ESE_SET_SPI_PWR           _IOW(NXP_ESE_MAGIC, 0x04, long)
 /* SPI or DWP can call this ioctl to get the current
  * power state of P61
 */
-#define PN8XT_ESE_GET_SPI_STATUS        _IOR(NXP_ESE_MAGIC, 0x05, unsigned int)
-#define PN8XT_ESE_GET_ACCESS            _IOW(NXP_ESE_MAGIC, 0x07, unsigned int)
-#define PN8XT_ESE_SET_PWR_SCM           _IOW(NXP_ESE_MAGIC, 0x08, unsigned int)
-#define PN8XT_ESE_SET_DN_STATUS         _IOW(NXP_ESE_MAGIC, 0x09, unsigned int)
-#define PN8XT_ESE_INHIBIT_PWR_CNTRL     _IOW(NXP_ESE_MAGIC, 0x0A, unsigned int)
+#define PN8XT_ESE_GET_SPI_STATUS        _IOR(NXP_ESE_MAGIC, 0x05, long)
+#define PN8XT_ESE_GET_ACCESS            _IOW(NXP_ESE_MAGIC, 0x07, long)
+#define PN8XT_ESE_SET_PWR_SCM           _IOW(NXP_ESE_MAGIC, 0x08, long)
+#define PN8XT_ESE_SET_DN_STATUS         _IOW(NXP_ESE_MAGIC, 0x09, long)
+#define PN8XT_ESE_INHIBIT_PWR_CNTRL     _IOW(NXP_ESE_MAGIC, 0x0A, long)
 
 long pn8xt_ese_dev_ioctl(struct ese_dev *ese_dev, unsigned int cmd, unsigned long arg);
 void pn8xt_ese_dev_open(struct ese_dev *ese_dev);
